BuyService  상품구매- 서영기

+++ src/main/java/com/green/babymeal/Buy/model/BuyInsDto.java
@@ -10,7 +10,7 @@ public class BuyInsDto {
     private String addressDetail;
     private String phoneNm;
     private String request;
-    private int payment;
+    private Byte payment;
     private int point;
 
     List<BuyInsOrderbasketDto> Insorderbasket;
